+++ Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.8.2026
+
+Marjor version update, version 0.18.26
+
+Changes:
+
+- Added Input and Output components.
+- Added firs working version of core business logic for components.
+
 ## 8.6.2026
 
 Marjor changes implemented, changes include: